(refactor)enum폴더 구조 변경
