+++ src/main/java/com/zipline/entity/enums/Platform.java
@@ -1,4 +1,4 @@
-package com.zipline.entity.publicItem.enums;
+package com.zipline.entity.enums;
 
 public enum Platform {
     NAVER,
