+++ src/main/java/com/zipline/service/publicItem/NaverRawArticleService.java
@@ -13,8 +13,8 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zipline.entity.publicItem.NaverRawArticle;
 import com.zipline.entity.publicItem.Region;
-import com.zipline.entity.publicItem.enums.CrawlStatus;
-import com.zipline.entity.publicItem.enums.MigrationStatus;
+import com.zipline.entity.enums.CrawlStatus;
+import com.zipline.entity.enums.MigrationStatus;
 import com.zipline.global.util.CoordinateUtil;
 import com.zipline.global.util.RandomSleepUtil;
 import com.zipline.repository.publicItem.NaverRawArticleRepository;
