+++ src/main/java/com/zipline/service/publicItem/ProxyNaverArticleService.java
@@ -25,8 +25,8 @@ import com.zipline.dto.publicItem.PageResultDTO;
 import com.zipline.dto.publicItem.ProxyInfoDTO;
 import com.zipline.entity.publicItem.PropertyArticle;
 import com.zipline.entity.publicItem.Region;
-import com.zipline.entity.publicItem.enums.CrawlStatus;
-import com.zipline.entity.publicItem.enums.Platform;
+import com.zipline.entity.enums.CrawlStatus;
+import com.zipline.entity.enums.Platform;
 import com.zipline.global.util.CoordinateUtil;
 import com.zipline.global.util.ProxyPool;
 import com.zipline.global.util.RandomSleepUtil;
