+++ src/main/java/com/zipline/service/publicItem/ProxyNaverRawArticleService.java
@@ -26,8 +26,8 @@ import com.zipline.dto.publicItem.PageResultDTO;
 import com.zipline.dto.publicItem.ProxyInfoDTO;
 import com.zipline.entity.publicItem.NaverRawArticle;
 import com.zipline.entity.publicItem.Region;
-import com.zipline.entity.publicItem.enums.CrawlStatus;
-import com.zipline.entity.publicItem.enums.MigrationStatus;
+import com.zipline.entity.enums.CrawlStatus;
+import com.zipline.entity.enums.MigrationStatus;
 import com.zipline.global.util.CoordinateUtil;
 import com.zipline.global.util.ProxyPool;
 import com.zipline.global.util.RandomSleepUtil;
