+++ src/main/java/com/zipline/service/publicItem/RegionCodeService.java
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zipline.dto.publicItem.RegionDTO;
 import com.zipline.entity.publicItem.Region;
-import com.zipline.entity.publicItem.enums.CrawlStatus;
+import com.zipline.entity.enums.CrawlStatus;
 import com.zipline.global.util.RandomSleepUtil;
 import com.zipline.repository.publicItem.RegionRepository;
